Use more caching when migrating information between document updates (#4414)
### Changes

Small improvement to performance by doing more caching when migrating
information between document updates.

Specifically, the cache `Dictionary<TextDocumentContentChangeEvent,
Position> getPositionAtEndOfAppliedChangeCache;` is re-used when
migrating different ranges/positions/verification trees/symbol tables

### Testing

No performance tests have been added/changes. We currently don't have
any.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>
