+++ Source/DafnyLanguageServer/Workspace/LanguageServerExtensions.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
         .AddSingleton<CreateProjectManager>(provider => (boogieEngine, documentIdentifier) => new ProjectManager(
           provider.GetRequiredService<DafnyOptions>(),
           provider.GetRequiredService<ILogger<ProjectManager>>(),
-          provider.GetRequiredService<IRelocator>(),
+          provider.GetRequiredService<CreateMigrator>(),
           provider.GetRequiredService<IFileSystem>(),
           provider.GetRequiredService<INotificationPublisher>(),
           provider.GetRequiredService<IVerificationProgressReporter>(),
@@ -48,7 +48,10 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
         })
         .AddSingleton<ITextDocumentLoader>(CreateTextDocumentLoader)
         .AddSingleton<INotificationPublisher, NotificationPublisher>()
-        .AddSingleton<IRelocator, Relocator>()
+        .AddSingleton<CreateMigrator>(provider => (changes, cancellationToken) => new Migrator(
+          provider.GetRequiredService<ILogger<Migrator>>(),
+          provider.GetRequiredService<ILogger<SignatureAndCompletionTable>>(),
+          changes, cancellationToken))
         .AddSingleton<ISymbolGuesser, SymbolGuesser>()
         .AddSingleton<ICompilationStatusNotificationPublisher, CompilationStatusNotificationPublisher>()
         .AddSingleton<ITelemetryPublisher, TelemetryPublisher>();
